Introduce CSS/SASS for the module, and add holidays
This commit introduces styling for the module. In the reality of the
module itself, I'll be changing this significantly to only style the
tables so that it can be used on other implementations in the future.
Currently, this is a personal project so it's not really ready for
public release yet.

Signed-off-by: Dave Russell Jr <dave@createazure.com>